Deduplicate TZ-forcing test boilerplate; fix unix-timestamp docstring
Extract the repeated TZ=Europe/Warsaw + time.tzset() setup/cleanup from
the four timestamp regression tests into a shared tests/tzutil.py
force_tz() helper, and correct human_timestamp_to_unix_timestamp()'s
docstring, which said the return type was float while the function
returns int.

+++ parsedmarc/utils.py
@@ -398,7 +398,7 @@ def human_timestamp_to_unix_timestamp(
             UTC wall-clock time instead of local time
 
     Returns:
-        float: The converted timestamp
+        int: The converted timestamp
     """
     human_timestamp = human_timestamp.replace("T", " ")
     return int(
